+++ Express-Course/.env
@@ -1,0 +1,1 @@
+dirname = D:\Carpeta para de cosas que si quiero\no quitar\VSCode\Repositorios\WebCourses\Express-Course\
